feat(tabs module): moved web home component into lazy tabs with shared mobile
