+++ src/app/app-routing.module.ts
@@ -1,9 +1,22 @@
 import { NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
-import { routes } from './app.routes';
+import { Routes, RouterModule } from '@angular/router';
+// import { routes } from './app.routes';
+
+export const routes: Routes = [
+  {
+      path: '',
+      redirectTo: '/tabs/home',
+      pathMatch: 'full',
+  },
+  {
+      path: 'tabs',
+      loadChildren: './tabs/tabs.module#TabsModule'
+  },
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
   exports: [RouterModule]
 })
+
 export class AppRoutingModule { }
