+++ src/app/app.module.ts
@@ -4,7 +4,6 @@ import { HttpClientModule, HttpClient } from '@angular/common/http';
 
 import { AppRoutingModule, } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { HomeComponent } from './home/home.component';
 import { StoreModule, ActionsSubject } from '@ngrx/store';
 import { CMSActionsSubject } from './services/dispatcher.service';
 import { reducers, metaReducers } from './reducers';
@@ -17,7 +16,6 @@ import { RecipeEffects } from './effects/recipes.effects';
 @NgModule({
   declarations: [
     AppComponent,
-    HomeComponent,
   ],
   imports: [
     BrowserModule,
