--- src/app/app.routes.ts
@@ -1,15 +1,0 @@
-import { Routes } from '@angular/router';
-
-import { HomeComponent } from './home/home.component';
-
-export const routes: Routes = [
-  {
-      path: '',
-      redirectTo: '/home',
-      pathMatch: 'full',
-  },
-  {
-      path: 'home',
-      component: HomeComponent,
-  },
-];
