--- src/app/home/home.component.tns.html
@@ -1,5 +1,0 @@
-<!--The content below is only a placeholder and can be replaced.-->
-<ActionBar [title]="title"></ActionBar>
-<StackLayout  class="p-20">
-    <Label text="Welcome to {{ title }}, Peter John!!" class="h1 text-center" textWrap="true"></Label>
-</StackLayout>
