+++ src/app/tabs/home/home.component.html
@@ -1,3 +1,16 @@
-<p>
-  home works!
-</p>
+<h1>Primary Categories!</h1>
+<div *ngFor="let category of categories$ | async">
+  <button (click)="getCategoryTypes(category.id)">{{category.name}}</button>
+</div>
+
+<h1>Secondary Categories</h1>
+<div *ngIf="subCategories$ | async as categoryTypes">
+  <div *ngFor="let subCategory of categoryTypes.subCategories">
+    <button (click)="selectSubCategory(subCategory.id)">{{subCategory.name}}</button>
+  </div>
+</div>
+
+<h1>Recipes!</h1>
+<div *ngFor="let recipe of recipes$ | async ">
+  <button>{{recipe.name}}</button>
+</div>
