+++ src/app/tabs/tabs.component.html
@@ -1,3 +1,1 @@
-<p>
-  tabs works!
-</p>
+<app-home></app-home>
